Use events for LogProxy, get rid of PromtailError as Exception
Send events like LokiPushApi already does so the UX is the same.

Instead of raising exceptions for everything in the LogProxy code,
forcing consumers to catch them to set state, fire back an event
if something happens with the Promtail Binary.

Only send a new `ContainerNotFound` Exception if the container name
is unset or ambiguous enough that we can't suss one out.

Add a bunch of type hints.

Update tests.

Closes: #70