Bring loki tests in line, too
While we're at it, unpin the operator hash.

Fix the unit test mocking to reflect the actual charm flow.

Emit the configure() hook when checking whether or not it can
connect or not.

Most importantly, FIX _provide_loki so it doesn't try to bring
up a second Provider if one is already set, otherwise
`ops.framework` sees two objects with the same handle and throws
an exception back at us.