A workload free Loki Tester Charm
This commit introduces a Loki tester charm for the purpose
of integration testing the Loki charm. The loki tester
charm forwards its own Python debug logs to Loki, when
a Juju relation is created between it and the Loki charm.
Normally the tester charm only forwards Python debug logs
to Loki. The tester charm also provides and Juju action
"log-error" which logs and error message to loki. The
tester charm contains an alert rule which is also forwarded
to Loki. This alert rule triggers an alert any time
an error log is received by Loki.