Update alert rules to not use "juju_unit" as part of the topology (#119)
When using the loki_push_api library, "juju_unit" is added to the alert rules. This causes the rule to only apply to the leader and not to the other unit. I used the updated JujuTopology class copied from prometheus_scrape with a minor modification to make it work in loki_push_api. This removes the "juju_unit" key from the alert rules.