Integration test updates
Add a helper method to also check the Prometheus-compatible
alerts endpoint from Loki itself, so we can check there as a
bisection point between Alertmanager.

Add a test which exercises this endpoint.

Typos and cleanup.

Add a (very temporary) FIXME so the "other" test which uses
Alertmananager can be moved to the bundle when this merges