Update consumers on pebble ready
This commit ensures that LokiPushApi consumers are updated and any new
alert rules set in the provider each time a pebble ready event is
triggered for the Provider. This is done because there is no gauranteed
ordering of pebble ready and relation changed events. Hence it can and
does happen that sometimes relation changed events are triggered for the
provider when it's workload container is not yet ready. To deal with
this situation a pebble ready event handler has been added that does
exactly the same thing as a relation changed event handler. Since the
workload container is only restarted when the layer configuration
changes but not when alert rule files change this change should not lead
to a infinite loop of container restarts. Also a guard has been added to
the relation chagned handler to check if pebble is ready. If pebble is
not ready in the relation changed handler the Provider charm does
nothing, since subsequent a pebble ready event is going to eventually
ensure that consumers and providers are in sync. This approach avoids
using the event.defer() method which can potentially lead to
unexpected behaviour.

This commit also adds a unit test to check that new alert rules are
updated on pebble ready.