Alertmanager Integration (#129)
* A workload free Loki Tester Charm

This commit introduces a Loki tester charm for the purpose
of integration testing the Loki charm. The loki tester
charm forwards its own Python debug logs to Loki, when
a Juju relation is created between it and the Loki charm.
Normally the tester charm only forwards Python debug logs
to Loki. The tester charm also provides and Juju action
"log-error" which logs and error message to loki. The
tester charm contains an alert rule which is also forwarded
to Loki. This alert rule triggers an alert any time
an error log is received by Loki.

* An alerting integration test

This commit adds an integration test to check that
LokiPushApiConsumer can forward logs and alert rules
to LokiPushApiProvider. The test further checks
conditions satisfying the alert rules do indeed raise
alerts.

* Clean up all the linter warnings and add integration tests for alertmanager

* Fix attribute error due to typo in method name

* Integration test updates

Add a helper method to also check the Prometheus-compatible
alerts endpoint from Loki itself, so we can check there as a
bisection point between Alertmanager.

Add a test which exercises this endpoint.

Typos and cleanup.

Add a (very temporary) FIXME so the "other" test which uses
Alertmananager can be moved to the bundle when this merges

* Make the linters happy

Co-authored-by: Balbir Thomas <balbir.thomas@canonical.com>