Instantiate Loki Provider only when ready
This commit ensures that the LokiPushApiProvider object is
instantiated only when Loki server is running and responding
to HTTP API requests. Failing to ensure this had the potential
of producing false negatives in the _check_alert_rules() method.
More significantly if the LokiPushApiProvider object should
not be forwarding its Push API URL to clients if it is not ready
to service HTTP requests.