Revert to an older pytest-operator
This commit reverts to an older pytest operator because the
current version of pytest operator is causing integration
tests to fail. This is happening because execeptions raised
when juju models used in integration tests which were
previously ignored are no longer silenced. It may be pertinent
to note that all tests in these models are passing but the
teardown process of the model leads to an error.