Remove tenacity, squelch urllib3 logging
Rename `_provide_loki` to `_loki_ready`, since it does not actually
have any involvement in starting the Loki server itself, and move
that check to `_configure` rather than `__init__` since its actual
readiness value has no bearing on the status of the provider
(which does talk to the workload container as a different issue,
and try to check the Loki API if those guards pass and alert rules
change, but both have properly caught exceptions).

As part and parcel of this. `urllib3` (as the parent module for
`requests`) and `requests` itself both inherit any given logging
level if they are imported, so setting the charm to `logging.DEBUG`
also flips those.

Turn `urllib3` into `logging.CRITICAL` only.

With these moved, it's possible to remove tenacity completely.